+++ kernels/openvx-core/c66x/bam/vx_bam_dilate3x3_target.c
@@ -334,8 +334,8 @@ static vx_status VX_CALLBACK tivxKernelDilateCreateInBamGraph(
         {
             memset(prms, 0, sizeof(tivxDilateParams));
 
-            node_list[*bam_node_cnt].nodeIndex = *bam_node_cnt;
-            node_list[*bam_node_cnt].kernelId = BAM_KERNELID_VXLIB_DILATE_3X3_I8U_O8U;
+            node_list[*bam_node_cnt].nodeIndex = (uint8_t)*bam_node_cnt;
+            node_list[*bam_node_cnt].kernelId = (uint32_t)BAM_KERNELID_VXLIB_DILATE_3X3_I8U_O8U;
             node_list[*bam_node_cnt].kernelArgs = NULL;
 
             BAM_VXLIB_dilate_3x3_i8u_o8u_getKernelInfo(NULL,
@@ -343,7 +343,7 @@ static vx_status VX_CALLBACK tivxKernelDilateCreateInBamGraph(
 
             kernel_details[*bam_node_cnt].compute_kernel_params = NULL;
 
-            prms->bam_node_num = *bam_node_cnt;
+            prms->bam_node_num = (uint8_t)*bam_node_cnt;
         }
         else
         {
@@ -384,11 +384,11 @@ static vx_status VX_CALLBACK tivxKernelDilateGetNodePort(
         {
             case TIVX_KERNEL_DILATE3X3_INPUT_IDX:
                 *bam_node = prms->bam_node_num;
-                *bam_port = BAM_VXLIB_DILATE_3X3_I8U_O8U_INPUT_IMAGE_PORT;
+                *bam_port = (uint8_t)BAM_VXLIB_DILATE_3X3_I8U_O8U_INPUT_IMAGE_PORT;
                 break;
             case TIVX_KERNEL_DILATE3X3_OUTPUT_IDX:
                 *bam_node = prms->bam_node_num;
-                *bam_port = BAM_VXLIB_DILATE_3X3_I8U_O8U_OUTPUT_IMAGE_PORT;
+                *bam_port = (uint8_t)BAM_VXLIB_DILATE_3X3_I8U_O8U_OUTPUT_IMAGE_PORT;
                 break;
             default:
                 VX_PRINT(VX_ZONE_ERROR,"tivxKernelDilateGetNodePort: non existing index queried by tivxKernelSupernodeCreate.tivxGetNodePort()\n");
