+++ kernels/openvx-core/c66x/bam/vx_bam_erode3x3_target.c
@@ -335,8 +335,8 @@ static vx_status VX_CALLBACK tivxKernelErodeCreateInBamGraph(
         {
             memset(prms, 0, sizeof(tivxErodeParams));
 
-            node_list[*bam_node_cnt].nodeIndex = *bam_node_cnt;
-            node_list[*bam_node_cnt].kernelId = BAM_KERNELID_VXLIB_ERODE_3X3_I8U_O8U;
+            node_list[*bam_node_cnt].nodeIndex = (uint8_t)*bam_node_cnt;
+            node_list[*bam_node_cnt].kernelId = (uint32_t)BAM_KERNELID_VXLIB_ERODE_3X3_I8U_O8U;
             node_list[*bam_node_cnt].kernelArgs = NULL;
 
             BAM_VXLIB_erode_3x3_i8u_o8u_getKernelInfo(NULL,
@@ -344,7 +344,7 @@ static vx_status VX_CALLBACK tivxKernelErodeCreateInBamGraph(
 
             kernel_details[*bam_node_cnt].compute_kernel_params = NULL;
 
-            prms->bam_node_num = *bam_node_cnt;
+            prms->bam_node_num = (uint8_t)*bam_node_cnt;
         }
         else
         {
@@ -385,11 +385,11 @@ static vx_status VX_CALLBACK tivxKernelErodeGetNodePort(
         {
             case TIVX_KERNEL_ERODE3X3_INPUT_IDX:
                 *bam_node = prms->bam_node_num;
-                *bam_port = BAM_VXLIB_ERODE_3X3_I8U_O8U_INPUT_IMAGE_PORT;
+                *bam_port = (uint8_t)BAM_VXLIB_ERODE_3X3_I8U_O8U_INPUT_IMAGE_PORT;
                 break;
             case TIVX_KERNEL_ERODE3X3_OUTPUT_IDX:
                 *bam_node = prms->bam_node_num;
-                *bam_port = BAM_VXLIB_ERODE_3X3_I8U_O8U_OUTPUT_IMAGE_PORT;
+                *bam_port = (uint8_t)BAM_VXLIB_ERODE_3X3_I8U_O8U_OUTPUT_IMAGE_PORT;
                 break;
             default:
                 VX_PRINT(VX_ZONE_ERROR,"tivxKernelErodeGetNodePort: non existing index queried by tivxKernelSupernodeCreate.tivxGetNodePort()\n");
