+++ kernels/openvx-core/c66x/bam/vx_bam_gaussian3x3_target.c
@@ -337,8 +337,8 @@ static vx_status VX_CALLBACK tivxKernelGaussianCreateInBamGraph(
         {
             memset(prms, 0, sizeof(tivxGaussianParams));
 
-            node_list[*bam_node_cnt].nodeIndex = *bam_node_cnt;
-            node_list[*bam_node_cnt].kernelId = BAM_KERNELID_VXLIB_GAUSSIAN_3X3_I8U_O8U;
+            node_list[*bam_node_cnt].nodeIndex = (uint8_t)*bam_node_cnt;
+            node_list[*bam_node_cnt].kernelId = (uint32_t)BAM_KERNELID_VXLIB_GAUSSIAN_3X3_I8U_O8U;
             node_list[*bam_node_cnt].kernelArgs = NULL;
 
             BAM_VXLIB_gaussian_3x3_i8u_o8u_getKernelInfo(NULL,
@@ -346,7 +346,7 @@ static vx_status VX_CALLBACK tivxKernelGaussianCreateInBamGraph(
 
             kernel_details[*bam_node_cnt].compute_kernel_params = NULL;
 
-            prms->bam_node_num = *bam_node_cnt;
+            prms->bam_node_num = (uint8_t)*bam_node_cnt;
         }
         else
         {
@@ -387,11 +387,11 @@ static vx_status VX_CALLBACK tivxKernelGaussianGetNodePort(
         {
             case TIVX_KERNEL_GAUSSIAN3X3_INPUT_IDX:
                 *bam_node = prms->bam_node_num;
-                *bam_port = BAM_VXLIB_GAUSSIAN_3X3_I8U_O8U_INPUT_IMAGE_PORT;
+                *bam_port = (uint8_t)BAM_VXLIB_GAUSSIAN_3X3_I8U_O8U_INPUT_IMAGE_PORT;
                 break;
             case TIVX_KERNEL_GAUSSIAN3X3_OUTPUT_IDX:
                 *bam_node = prms->bam_node_num;
-                *bam_port = BAM_VXLIB_GAUSSIAN_3X3_I8U_O8U_OUTPUT_IMAGE_PORT;
+                *bam_port = (uint8_t)BAM_VXLIB_GAUSSIAN_3X3_I8U_O8U_OUTPUT_IMAGE_PORT;
                 break;
             default:
                 VX_PRINT(VX_ZONE_ERROR,"tivxKernelGaussianGetNodePort: non existing index queried by tivxKernelSupernodeCreate.tivxGetNodePort()\n");
